+++ NEWS.md
@@ -7,6 +7,8 @@
 
 * Font family names are now quoted (#136)
 
+* svglite now renders leading white-space in <text> nodes. (#131, @hrbrmstr)
+
 # svglite 2.0.0
 
 * svglite now uses systemfonts for text metric calculations and font family